requirements.txt
pandas>=2.0
openpyxl>=3.1
requests>=2.28
pyyaml>=6.0
